+++ dashboard/constants.py
@@ -31,4 +31,14 @@ DB_ENV_VARS = ('DATABASE_NAME', 'DATABASE_USER', 'DATABASE_PASSWD',
                'DATABASE_HOST', 'TRANSIFEX_USER', 'TRANSIFEX_PASSWD')
 
 # Job Types
-TS_JOB_TYPES = ('synctransplatform', 'syncrelschedule', 'syncupstream')
+TS_JOB_TYPES = ('synctransplatform', 'syncrelschedule', 'syncupstream',
+                'syncdownstream', 'syncbuildtags')
+
+# Branch Mapping Keys
+BRANCH_MAPPING_KEYS = ('platform_version', 'buildsys', 'buildsys_tag', 'upstream_release')
+
+# Supported Build Systems
+BUILD_SYSTEMS = ('brew', 'koji')
+
+# YML Jobs Execution Types
+JOB_EXEC_TYPES = ('sequential', 'parallel')